UK: move reset password to ConnectViewSet
